ai_service/requirements.txt
# AI/ML and Image Processing
flask==3.0.0
numpy==1.25.2
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10

# Natural Language Processing
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99

# HTTP and API
requests==2.31.0
flask-cors==4.0.0

# Data processing
pandas==2.1.3

# Audio processing (for future voice features)
speechrecognition==3.10.0
